Solution to Streaming Service Subscription Management

+++ Streaming Service Subscription Management@@67877d61dc107486ad5c2752/code.c
@@ -6,7 +6,7 @@
 struct Subscription {
     char userName[50];
     char subType[10];
-    int cost;
+    float cost;
 };
 
 int main() {
@@ -19,7 +19,7 @@ int main() {
     float Revenue1 = 0, Revenue2 = 0, Revenue3 = 0;
 
     for (int i = 0; i < N; i++) {
-        scanf("%s %s %d", users[i].userName, users[i].subType, &users[i].cost);
+        scanf("%s %s %f", users[i].userName, users[i].subType, &users[i].cost);
 
         if (strcmp(users[i].subType, "Basic") == 0) {
             Count1++;
